[REFACTOR] implement separatioof concern dengan memisahkan schemas dari ke filenya sendiri
